reward should not be int

+++ buffer.py
@@ -17,7 +17,7 @@ class PrioritizedReplayBuffer:
         # transition: state, action, reward, next_state, done
         self.state = torch.empty(buffer_size, 4, 208, 300, dtype=torch.uint8)
         self.action = torch.empty(buffer_size, action_size, dtype=torch.int)
-        self.reward = torch.empty(buffer_size, dtype=torch.int)
+        self.reward = torch.empty(buffer_size, dtype=torch.float32)
         self.next_state = torch.empty(buffer_size, 4, 208, 300, dtype=torch.uint8)
         self.done = torch.empty(buffer_size, dtype=torch.int)
